fixed resp time < 1ms

+++ backend/models/ping.go
@@ -6,6 +6,6 @@ type PingResult struct {
 	ID            int       `json:"id,omitempty"`
 	IPAddress     string    `json:"ip_address"`
 	ContainerName string    `json:"container_name"`
-	PingTime      int       `json:"ping_time"`
+	PingTime      float64   `json:"ping_time"`
 	LastSuccess   time.Time `json:"last_success"`
 }
